+++ src/components/header/header_links/header_links.tsx
@@ -17,83 +17,140 @@
  * under the License.
  */
 
-import React, { Component } from 'react';
+import React, {
+  HTMLAttributes,
+  FunctionComponent,
+  useState,
+  useEffect,
+  MouseEventHandler,
+} from 'react';
 import classNames from 'classnames';
 
-import { CommonProps } from '../../common';
-import { EuiIcon } from '../../icon';
-import { EuiPopover } from '../../popover';
+import { CommonProps, keysOf } from '../../common';
+import { EuiIcon, IconType } from '../../icon';
+import { EuiPopover, EuiPopoverProps } from '../../popover';
 import { EuiI18n } from '../../i18n';
-import { EuiHeaderSectionItemButton } from '../header_section';
+import {
+  EuiHeaderSectionItemButton,
+  EuiHeaderSectionItemButtonProps,
+} from '../header_section';
+import { EuiBreakpointSize } from '../../../services/breakpoint';
+import { EuiHideFor, EuiShowFor } from '../../responsive';
 
-interface State {
-  isOpen: boolean;
-}
+type EuiHeaderLinksGutterSize = 'xs' | 's' | 'm' | 'l';
+type EuiHeaderLinksPopoverButtonProps = EuiHeaderSectionItemButtonProps & {
+  iconType?: IconType;
+};
 
-export class EuiHeaderLinks extends Component<CommonProps, State> {
-  state: State = {
-    isOpen: false,
+export type EuiHeaderLinksProps = CommonProps &
+  HTMLAttributes<HTMLElement> & {
+    /**
+     * Spacing between direct children
+     */
+    gutterSize?: EuiHeaderLinksGutterSize;
+    /**
+     * A list of named breakpoints at which to show the popover version
+     */
+    popoverBreakpoints?: EuiBreakpointSize[] | 'all' | 'none';
+    /**
+     * Extend the functionality of the EuiPopover.button which is a EuiHeaderSectionItemButton.
+     * With the addition of `iconType` to change the display icon which defaults to `apps`
+     */
+    popoverButtonProps?: EuiHeaderLinksPopoverButtonProps;
+    /**
+     * Extend the functionality of the EuiPopover
+     */
+    popoverProps?: Omit<EuiPopoverProps, 'button' | 'closePopover'>;
   };
 
-  onMenuButtonClick = () => {
-    this.setState({
-      isOpen: !this.state.isOpen,
-    });
+const gutterSizeToClassNameMap: {
+  [gutterSize in EuiHeaderLinksGutterSize]: string;
+} = {
+  xs: '--gutterXS',
+  s: '--gutterS',
+  m: '--gutterM',
+  l: '--gutterL',
+};
+export const GUTTER_SIZES = keysOf(gutterSizeToClassNameMap);
+
+export const EuiHeaderLinks: FunctionComponent<EuiHeaderLinksProps> = ({
+  children,
+  className,
+  gutterSize = 's',
+  popoverBreakpoints = ['xs', 's'],
+  popoverButtonProps,
+  popoverProps,
+  ...rest
+}) => {
+  const { onClick: _onClick, iconType = 'apps', ...popoverButtonRest } = {
+    ...popoverButtonProps,
   };
 
-  closeMenu = () => {
-    this.setState({
-      isOpen: false,
-    });
+  const [mobileMenuIsOpen, setMobileMenuIsOpen] = useState(false);
+
+  const onMenuButtonClick: MouseEventHandler<HTMLButtonElement> = e => {
+    _onClick && _onClick(e);
+    setMobileMenuIsOpen(!mobileMenuIsOpen);
   };
 
-  componentDidMount() {
-    window.addEventListener('resize', this.closeMenu);
-  }
+  const closeMenu = () => {
+    setMobileMenuIsOpen(false);
+  };
 
-  componentWillUnmount() {
-    window.removeEventListener('resize', this.closeMenu);
-  }
+  useEffect(() => {
+    window.addEventListener('resize', closeMenu);
+    return () => {
+      window.removeEventListener('resize', closeMenu);
+    };
+  });
 
-  render() {
-    const { children, className, ...rest } = this.props;
+  const classes = classNames('euiHeaderLinks', className);
 
-    const classes = classNames('euiHeaderLinks', className);
+  const button = (
+    <EuiI18n token="euiHeaderLinks.openNavigationMenu" default="Open menu">
+      {(openNavigationMenu: string) => (
+        <EuiHeaderSectionItemButton
+          aria-label={openNavigationMenu}
+          onClick={onMenuButtonClick}
+          {...popoverButtonRest}>
+          <EuiIcon type={iconType} size="m" />
+        </EuiHeaderSectionItemButton>
+      )}
+    </EuiI18n>
+  );
 
-    const button = (
-      <EuiI18n
-        token="euiHeaderLinks.openNavigationMenu"
-        default="Open navigation menu">
-        {(openNavigationMenu: string) => (
-          <EuiHeaderSectionItemButton
-            aria-label={openNavigationMenu}
-            onClick={this.onMenuButtonClick}>
-            <EuiIcon type="apps" size="m" />
-          </EuiHeaderSectionItemButton>
-        )}
-      </EuiI18n>
-    );
+  return (
+    <EuiI18n token="euiHeaderLinks.appNavigation" default="App menu">
+      {(appNavigation: string) => (
+        <nav className={classes} aria-label={appNavigation} {...rest}>
+          <EuiHideFor sizes={popoverBreakpoints}>
+            <div
+              className={classNames('euiHeaderLinks__list', [
+                `euiHeaderLinks__list${gutterSizeToClassNameMap[gutterSize]}`,
+              ])}>
+              {children}
+            </div>
+          </EuiHideFor>
 
-    return (
-      <EuiI18n token="euiHeaderLinks.appNavigation" default="App navigation">
-        {(appNavigation: string) => (
-          <nav className={classes} aria-label={appNavigation} {...rest}>
-            <div className="euiHeaderLinks__list">{children}</div>
-
+          <EuiShowFor sizes={popoverBreakpoints}>
             <EuiPopover
-              className="euiHeaderLinks__mobile"
               ownFocus
               button={button}
-              isOpen={this.state.isOpen}
+              isOpen={mobileMenuIsOpen}
               anchorPosition="downRight"
-              closePopover={this.closeMenu}
-              panelClassName="euiHeaderLinks__mobileList"
-              panelPaddingSize="none">
-              {children}
+              closePopover={closeMenu}
+              panelPaddingSize="none"
+              {...popoverProps}>
+              <div
+                className={classNames('euiHeaderLinks__mobileList', [
+                  `euiHeaderLinks__mobileList${gutterSizeToClassNameMap[gutterSize]}`,
+                ])}>
+                {children}
+              </div>
             </EuiPopover>
-          </nav>
-        )}
-      </EuiI18n>
-    );
-  }
-}
+          </EuiShowFor>
+        </nav>
+      )}
+    </EuiI18n>
+  );
+};